% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mnu.R
\name{plot_mnu}
\alias{plot_mnu}
\title{Plot Mean Normalized Usage (MNU) results for terrain and imagery covariates}
\usage{
plot_mnu(df)
}
\arguments{
\item{df}{Output from `mean_normalized_usage()` function.}
}
\value{
A combined grid plot object.
}
\description{
Creates side-by-side ggplot2 bar charts showing MNU grouped by analysis scale
or temporal/statistical aggregation, faceted by depth class.
}
